requirement.txt
openai
langchain
streamlit
transformers
PyPDF2
faiss-cpu
tiktoken
unstructured
chromadb
PyMuPDF
pdf2image
pdfminer.six
opencv-python
tensorflow
tensorflow_hub
pdfplumber
list
faiss-gpu
chroma-hnswlib
pdfminer.six







